chore: fix test configuration (#53)
This makes a few changes to get tests passing. I think one of the problems we were facing with drone builds failing is that react-scripts expects CI=true to be set in CI environments. For some reason our drone doesn't so this adds that and changes the commands a bit to actually work. I also decided that we might as well just show coverage by default when running test. Objections?